+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@
 
 - Added base functionality for layout manager.
 - Adding base functionality for Horizontal and Vertical layout managers.
+- Added "add_widget_to_layout_manager" function.
+- Layout Manager redesign is working properly.
+- Created layout manager that is part of the widget store. (#150)
 
 ## 0.3.4